@@ -1,17 +1,17 @@
 ﻿using Competition.Validation;
 using Coordinates;
-using LoggerComponent;
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Text;
+using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace Competition
 {
     public class DonutTask : ICompetitionTask
     {
+        [JsonIgnore()]
+        private readonly ILogger<DonutTask> Logger;    
+
         #region Properties
         /// <summary>
         /// The task number
@@ -89,7 +89,7 @@ namespace Competition
         /// <para>mandatory</para>
         /// </summary>
         [JsonProperty("Is reentrance allowed")]
-        public bool IsReentranceAllowed
+        public bool IsReEntranceAllowed
         {
             get; set;
         } = true;
@@ -115,15 +115,13 @@ namespace Competition
         /// <returns>true:success;false:error</returns>
         public bool CalculateResults(Track track, bool useGPSAltitude, out double result)
         {
-            string functionErrorMessage = $"Failed to calculate result for {this} and Pilot '#{track.Pilot.PilotNumber}{(!string.IsNullOrWhiteSpace(track.Pilot.FirstName) ? $"({track.Pilot.FirstName},{track.Pilot.LastName})" : "")}': ";
             result = 0.0;
             List<(int trackPointNumber, Coordinate coordinate)> trackPointsInDonut = new List<(int trackPointNumber, Coordinate coordinate)>();
 
             Declaration targetDeclaration = ValidationHelper.GetValidDeclaration(track, GoalNumber, DeclarationValidationRules);
             if (targetDeclaration == null)
             {
-                //Debug.WriteLine("No valid goal found");
-                Log(LogSeverityType.Error, functionErrorMessage + $"No valid goal found for goal '#{GoalNumber}'");
+                Logger.LogError("Failed to calculate result for '{task}' and Pilot '#{pilotNumber}{pilotName}': No valid goal found for goal '#{goalNumber}'", ToString(), track.Pilot.PilotNumber, (!string.IsNullOrWhiteSpace(track.Pilot.FirstName) ? $"({track.Pilot.FirstName},{track.Pilot.LastName})" : ""), GoalNumber);
                 return false;
             }
             List<Coordinate> coordinates = track.TrackPoints;
@@ -174,7 +172,7 @@ namespace Competition
             }
 
 
-            if (!IsReentranceAllowed)//evaluate first chunk only
+            if (!IsReEntranceAllowed)//evaluate first chunk only
             {
                 if (chunksInDonut[0].Count >= 2)
                 {
@@ -214,9 +212,9 @@ namespace Competition
         /// <param name="outerRadius">The radius of the outer circle in meter (mandatory)</param>
         /// <param name="lowerBoundary">Lower boundary of the donut in meter (optional; use double.NaN to omit)</param>
         /// <param name="upperBoundary">Upper boundary of the donut in meter (optional; use double.NaN to omit)</param>
-        /// <param name="isReentranceAllowed">Specify whether or not re-entrance in the donut is allowed (mandatory)</param>
+        /// <param name="isReEntranceAllowed">Specify whether or not re-entrance in the donut is allowed (mandatory)</param>
         /// <param name="declarationValidationRules">List of rules for declaration validation (optional; leave list empty to omit)</param>
-        public void SetupDonut(int taskNumber, int goalNumber, int numberOfDeclarations, double innerRadius, double outerRadius, double lowerBoundary, double upperBoundary, bool isReentranceAllowed, List<IDeclarationValidationRules> declarationValidationRules)
+        public void SetupDonut(int taskNumber, int goalNumber, int numberOfDeclarations, double innerRadius, double outerRadius, double lowerBoundary, double upperBoundary, bool isReEntranceAllowed, List<IDeclarationValidationRules> declarationValidationRules)
         {
             TaskNumber = taskNumber;
             GoalNumber = goalNumber;
@@ -225,7 +223,7 @@ namespace Competition
             OuterRadius = outerRadius;
             LowerBoundary = lowerBoundary;
             UpperBoundary = upperBoundary;
-            IsReentranceAllowed = isReentranceAllowed;
+            IsReEntranceAllowed = isReEntranceAllowed;
             DeclarationValidationRules = declarationValidationRules;
         }
 
@@ -236,10 +234,6 @@ namespace Competition
         #endregion
 
         #region Private methods
-        private void Log(LogSeverityType logSeverity, string text)
-        {
-            Logger.Log(this, logSeverity, text);
-        }
         #endregion
     }
 }
