@@ -1,17 +1,19 @@
-﻿using Coordinates;
-using LoggerComponent;
+﻿using Competition.Validation;
+using Coordinates;
+using Microsoft.Extensions.Logging;
+using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Competition.Validation;
 
 namespace Competition
 {
     public class HesitationWaltzTask : ICompetitionTask
     {
         #region Properties
+        [JsonIgnore()]
+        private readonly ILogger<HesitationWaltzTask> Logger;
+
         public int TaskNumber
         {
             get;
@@ -55,7 +57,7 @@ namespace Competition
             result = double.NaN;
             if (!ValidationHelper.IsMarkerValid(track, MarkerNumber, MarkerValidationRules))
             {
-                Log(LogSeverityType.Error, functionErrorMessage + $"Marker '{MarkerNumber}' is invalid or doesn't exists");
+                Logger.LogError("Failed to calculate result for '{task}' and Pilot '#{pilotNumber}{pilotName}': Marker '{markerNumber}' is invalid or doesn't exists", ToString(), track.Pilot.PilotNumber, (!string.IsNullOrWhiteSpace(track.Pilot.FirstName) ? $"({track.Pilot.FirstName},{track.Pilot.LastName})" : ""), MarkerNumber);
                 return false;
             }
 
@@ -72,13 +74,13 @@ namespace Competition
                         Goals = CalculateGoals(track);
                         if (Goals.Count == 0)
                         {
-                            Log(LogSeverityType.Error, $"Failed to calculate goals: no goals could be calculated");
+                            Logger.LogError("Failed to calculate result for '{task}' and Pilot '#{pilotNumber}{pilotName}': No goals could be calculated", ToString(), track.Pilot.PilotNumber, (!string.IsNullOrWhiteSpace(track.Pilot.FirstName) ? $"({track.Pilot.FirstName},{track.Pilot.LastName})" : ""));
                             return false;
                         }
                     }
                     catch (Exception ex)
                     {
-                        Log(LogSeverityType.Error, $"Failed to calculate goals: {ex.Message}");
+                        Logger.LogError(ex,"Failed to calculate result for '{task}' and Pilot '#{pilotNumber}{pilotName}'", ToString(), track.Pilot.PilotNumber, (!string.IsNullOrWhiteSpace(track.Pilot.FirstName) ? $"({track.Pilot.FirstName},{track.Pilot.LastName})" : ""));
                         return false;
                     }
                 }
@@ -126,10 +128,6 @@ namespace Competition
         #endregion
 
         #region Private methods
-        private void Log(LogSeverityType logSeverity, string text)
-        {
-            Logger.Log(this, logSeverity, text);
-        }
         #endregion
     }
 }
