@@ -1,16 +1,15 @@
 ﻿using Coordinates;
-using LoggerComponent;
+using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Competition.Validation
 {
     public class MarkerToMarkerTimingRule : IMarkerValidationRules
     {
         #region Properties
+
+        private readonly ILogger<MarkerToMarkerTimingRule> Logger;
         /// <summary>
         /// Earliest time to drop second marker
         /// <para>optional; use double.NaN to omit</para>
@@ -87,11 +86,9 @@ namespace Competition.Validation
             return "Marker to Markers Timing Rule";
         }
         #endregion
+
         #region Private methods
-        private void Log(LogSeverityType logSeverity, string text)
-        {
-            Logger.Log(this, logSeverity, text);
-        }
+
         #endregion
     }
 }
